Move name resolution files and parsing files to their own folders
