+++ quint/src/index.ts
@@ -5,23 +5,23 @@
 export * from './quintTypes'
 export * from './quintIr'
 export * from './IRprinting'
-export * from './quintParserFrontend'
-export * from './definitionsCollector'
+export * from './parsing/quintParserFrontend'
+export * from './names/definitionsCollector'
 export * from './effects/base'
 export * from './effects/printing'
 export * from './effects/builtinSignatures'
 export { typeSchemeToString } from './types/printing'
 export * from './errorTree'
-export { LookupTable } from './lookupTable'
+export { LookupTable } from './names/lookupTable'
 export { produceDocsById, DocumentationEntry } from './docs'
 export { builtinDocs } from './builtin'
 export * as Builder from '../test/builders/ir'
 export { TypeScheme } from './types/base'
 export { findExpressionWithId, findTypeWithId, findDefinitionWithId } from './IRFinder'
-export * from './scoping'
+export * from './names/scoping'
 export * from './quintAnalyzer'
 export * from './quintError'
 export { newIdGenerator, IdGenerator } from './idGenerator'
-export { fileSourceResolver } from './sourceResolver'
+export { fileSourceResolver } from './parsing/sourceResolver'
 export { format } from './prettierimp'
 export { prettyQuintEx, prettyTypeScheme, prettyQuintDef } from './graphics'
