+++ quint/src/names/nameResolver.ts
@@ -15,11 +15,11 @@
  * @module
  */
 
-import { QuintApp, QuintDef, QuintInstance, QuintModule, QuintName } from './quintIr'
-import { QuintConstType } from './quintTypes'
+import { QuintApp, QuintDef, QuintInstance, QuintModule, QuintName } from '../quintIr'
+import { QuintConstType } from '../quintTypes'
 import { ScopeTree } from './scoping'
 import { LookupTable } from './lookupTable'
-import { IRVisitor, walkModule } from './IRVisitor'
+import { IRVisitor, walkModule } from '../IRVisitor'
 import { Either, left, right } from '@sweet-monads/either'
 import { DefinitionsByName, lookupType, lookupValue } from './definitionsByName'
