+++ quint/src/repl.ts
@@ -31,12 +31,12 @@ import {
 import { formatError } from './errorReporter'
 import { ComputableKind, EvalResult, Register, kindName } from './runtime/runtime'
 import { TraceRecorder, newTraceRecorder, noExecutionListener } from './runtime/trace'
-import { ErrorMessage, parseExpressionOrUnit } from './quintParserFrontend'
+import { ErrorMessage, parseExpressionOrUnit } from './parsing/quintParserFrontend'
 import { prettyQuintEx, printExecutionFrameRec, terminalWidth } from './graphics'
 import { verbosity } from './verbosity'
 import { newRng } from './rng'
 import { version } from './version'
-import { fileSourceResolver } from './sourceResolver'
+import { fileSourceResolver } from './parsing/sourceResolver'
 import { cwd } from 'process'
 import { newIdGenerator } from './idGenerator'
 import { definitionToString } from './IRprinting'
