+++ quint/src/types/substitutions.ts
@@ -14,7 +14,7 @@
 
 import { Either } from '@sweet-monads/either'
 import { ErrorTree, errorTreeToString } from '../errorTree'
-import { LookupTable } from '../lookupTable'
+import { LookupTable } from '../names/lookupTable'
 import { QuintType, Row } from '../quintTypes'
 import { Constraint } from './base'
 import { unify, unifyRows } from './constraintSolver'
