+++ quint/test/effects/modeChecker.test.ts
@@ -5,10 +5,10 @@ import { OpQualifier } from '../../src/quintIr'
 import { EffectInferrer } from '../../src/effects/inferrer'
 import { ModeChecker } from '../../src/effects/modeChecker'
 import { QuintError, quintErrorToString } from '../../src/quintError'
-import { resolveNames } from '../../src/nameResolver'
+import { resolveNames } from '../../src/names/nameResolver'
 import JSONbig from 'json-bigint'
-import { collectDefinitions } from '../../src/definitionsCollector'
-import { treeFromModule } from '../../src/scoping'
+import { collectDefinitions } from '../../src/names/definitionsCollector'
+import { treeFromModule } from '../../src/names/scoping'
 import { moduleToString } from '../../src/IRprinting'
 import { errorTreeToString } from '../../src/errorTree'
