+++ quint/test/types/constraintGenerator.test.ts
@@ -7,10 +7,10 @@ import { walkModule } from '../../src/IRVisitor'
 import { buildModuleWithDefs } from '../builders/ir'
 import { constraintToString } from '../../src/types/printing'
 import { ErrorTree } from '../../src/errorTree'
-import { collectDefinitions } from '../../src/definitionsCollector'
-import { LookupTable } from '../../src/lookupTable'
-import { resolveNames } from '../../src/nameResolver'
-import { treeFromModule } from '../../src/scoping'
+import { collectDefinitions } from '../../src/names/definitionsCollector'
+import { LookupTable } from '../../src/names/lookupTable'
+import { resolveNames } from '../../src/names/nameResolver'
+import { treeFromModule } from '../../src/names/scoping'
 import JSONbig from 'json-bigint'
 
 describe('ConstraintGeneratorVisitor', () => {
