+++ quint/test/types/constraintSolver.test.ts
@@ -7,7 +7,7 @@ import { substitutionsToString } from '../../src/types/printing'
 import { Substitutions } from '../../src/types/substitutions'
 import { Row } from '../../src/quintTypes'
 import { errorTreeToString } from '../../src/errorTree'
-import { LookupTable } from '../../src/lookupTable'
+import { LookupTable } from '../../src/names/lookupTable'
 
 const table: LookupTable = new Map([
   // A type alias (ref 1n)
